Linkcheck: Avoid duplicate download of page to check anchors
Anchors (named fragments in the spec) indicate a secondary resource
relative to the primary resource. The page can be fetched once, and all
anchors checked on the retrieved page.

The distinction between anchors and the web page location has been made
more explicit in the work queue. A single page now has a set of anchors
to check.
Previously, the anchor was considered part of the link, such that
http://a.example/#id-1 and http://a.example/#id-2 would be considered
two different links. After this change, the link becomes
http://a.example/ and has two anchors: {'id-1', 'id-2'}.

Errors occurring when the web page is downloaded no longer show the
anchor, the anchor existence is conditioned by the existence of the page
at URI.

Fix #4303