linkcheck: Skip links to ignore in post transform
Avoid passing the links to the checker threads when it should not be
checked.

The log files are now written by the link checker and the post
transform, share the file object.

The check results structure are now used by the link checker and the
post transform, formalize its structure in a named tuple.