Revert "Closes: #8807 by letting sphinxadmonition step a counter"
This reverts commit 16e3a931c47b47a5330817e09012641bcbbae756.

Indeed this may have some after-effect. In current situation, todo
directive issues latex `\label` but no reference anchor is set, so
reference via LaTeX `\ref` will link probably to sectioning unit. With
the reverted commit, the link will be to the notice/admonition but with
some numbered label not appearing otherwise in document.

Anyway, this issue is related to only small extend to the task of
achieving permanent named destinations in PDF.