LaTeX: partially revert #8997 \pysigline
Reason is that mark-up such as this:

   .. attribute:: state
                  state_machine

      The state and state machine which controls the parsing.  Used for
      ``nested_parse``.

generates two `\pysigline` each with `\phantomsection`. The latex code
to get good vertical spacing between label and its description, if label
``\parbox`` is multi-line, cause in this context the two (generally,
single-line) labels to be stacked vertically with no spacing.

This commit keeps the ``\parbox`` which fixes #8980, but drops the
attempt to correct vertical distance to description, so as to not alter
the possibly more common use case of items with common description.