Improve progress logging on parallel builds
Currently, the progress bar increases when each parallel task is submitted for
processing. However, that doesn't necessarily mean any progress has actually
been made yet. The result is the progress bar very quickly jumps to 100% and stalls at

> waiting for workers ...

This updates the progress bar as individual tasks are joined, for
long-running builds this creates a meaningfully sense of progress.