Fix #9224: info-field-list does not support a type containing space
So far, DocFieldTransformer has split the field name for `:type:`
and `:param:` at the first whitespace.  Therefore, it does not allow
to use a type containing whitespace (ex. `Dict[str, str]`).

This changes the splitting logic to divide the field name at the last
whitespace (a.k.a `rstrip()`).  It allows to use whitespaces on types.