Revert the removal of sphinx.util:force_decode()
After the release of 4.0.0, some 3rd party extensions have became not
working with the latest Sphinx because `force_decode()` function was
removed.  It was deprecated since Sphinx-2.0 and warned for the removal
since 3.0.

This reverts the removal and extends its deprecation period to 5.0.0.  I
hope it helps users of these extensions.