Fix rendering of nested type annotations with alias defined
The current implementation replaces type annotations with a alias defined by the
class TypeAliasForwardRef. This class is later again replaced by the type
defined in the alias map.

This commit fixes the problem that this second replacement does not happen if
the aliased type is nested in another generic
type (e.g. Iterator[aliased_type]). The new function goes recursivly through the
nested types and replaces the internal class TypeAliasForwardRef again by the
correct annotation.