LaTeX: fix '\raggedright' escaping causing "aggedright" text
Sphinx version 4.0 introduced a bug in handling hlists in its LaTeX backend.
Due to improper backslash escaping,
LaTeX "\raggedright" command gets written as
Carriage Return character (0x0D) followed by "aggedright".

This results in stray "aggedright" text appearing in the resulting PDF
and lack of effect \raggedright was supposed to achieve.

Fix this by converting the remaining string to a raw string.

This appears to be the only occurrence of such a missing escaping
based on a quick grep.

Fixes #9734.

Fixes: 20884bb0c9f7: "refactor: LaTeX: Use raw strings for LaTeX macros"