Make util.typing.restify sanitise unreproducible output (eg. memory addresses)
Whilst working on the Reproducible Builds effort [0] I noticed that sphinx
generates output that is not reproducible, causing a number of packages in
Debian to unreproducible.

Specifically, when Sphinx locates an alias of an instance when generating
'autodoc' documentation, it uses the raw Python repr(...) of the object and
does not sanitise it for memory addresses (etc.) like elsewhere in Sphinx.

This can result in documentation like this:

-<dd><p>alias of &lt;webob.client.SendRequest object at 0x7fd769189df0&gt;</p>
+<dd><p>alias of &lt;webob.client.SendRequest object at 0x7f0f02233df0&gt;</p>

Patch attached that uses the object_description method, which was added to fix
precisely this kind of issue.

I originally filed this in Debian as bug #996948 [1].

 [0] https://reproducible-builds.org/
 [1] https://bugs.debian.org/996948