nodes: support not replacing inlined toctree nodes
The `inline_all_toctrees` will replace toctree nodes with the contents
of each toctree entry. While replacing the toctree node is preferable
for most builders (and themes), some uses cases may wish to leave the
TOC tree in. For example, if an HTML theme does not support a sidebar
which produces a TOC tree, the resulting documentation will not have any
TOC tree in the final output.

This commit introduces a `replace` option to the `inline_all_toctrees`
function which allows a builder to explicitly decide if toctree nodes
are replaced (the previous use case and the default state) or leave the
toctree nodes in (which can later be resolved on).

Signed-off-by: James Knight <james.d.knight@live.com>