singlehtml: option to include toc tree entries in the main document
Documentation which use an HTML theme which do not support a table of
contents (e.g. `basic`) or a configuration which has `nosidebar` set,
there will be no TOC tree generated in the final document. To support
the use case where a user wishes to have TOC tree entries rendered in
the document, a `singlehtml_toctree` option can be set to `True` to
enable this.

Signed-off-by: James Knight <james.d.knight@live.com>