Fix #10058: autosummary: Imported members are not shown
Originally, the `check_module()` call was added at 21b8384 to hide
imported members on generating stubs.  But it was incorrect approach.
Therefore it was disabled by b433197 and fixed the original issue by
another approach at 2390c554.

Finally, the `check_module()` call becomes meaningless code.  But, at
present, it causes that imported members are not shown when
`autodoc_class_signature` is 'separated'.  To resolve the problem, this
removes the meaningless call.