LaTeX: comment out footnote mark enhancement for footnote environment
Page detection left to apply as was case before this PR, i.e. only for
the \sphinxfootnotemark mark-up, which means either in combination
with footnotetext environment, or for multiply referred-to footnotes
and the extra footnote marks additional to the initial one.

This is to avoid a problem with each LaTeX pass in certain circumstances
modifying the aux file, and latexmk consequently complaining.

refs:
https://github.com/sphinx-doc/sphinx/pull/10191#issuecomment-1038807448