config: Add 'link_inline_targets'
In 4.4.0, we started allowing users to reference an inline target via
the 'ref' role. This had the unfortunate side-effect of changing inline
targets from locally scoped to globally scoped, breaking many document
trees in the process. Add a new config option, 'link_inline_targets',
which will allow users to opt-in to this behavior if they wish to use
it. It is disabled by default.

Signed-off-by: Stephen Finucane <stephen@that.guru>