Factor out common logic from HTMLTranslator and HTML5Translator
Previously, HTML5Translator contained a lot of code copied from HTMLTranslator,
which created a maintenance burden and made it difficult to know what
the actual differences are.

With this change the common logic is factored out into a mixin, which
eliminates the code duplication.

This change also elimintes the `generate_targets_for_listing` fix from
HTMLTranslator now that Sphinx depends on a version of docutils that
contains the fix.