logging: always show source locations as absolute paths
Nodes attached to the parse tree via the `include` directive have
their `source` attribute set to the relative path of the included
file. Other nodes in the tree use the full absolute path. For
consistent logging, ensure that the node location is always expressed
as an absolute path, when the filename is known.

See https://github.com/sphinx-contrib/spelling/issues/153 for an
example of the effect of the original problem.