ObjectDescription: Do not create a target node without ids
It might be that a object description has no ids associated with it,
e.g. if the option :noindex: is passed.  In this case we would replace
the object description node with a target node, which has no ids.  This
breaks docutils assumption about a target node and leads not errors.
Therefore, do only create a target node if we have ids to work with.