Add CITATION.cff
Add CITATION.cff file. Closes #10543.

I have tried to add all public members of the sphinx org as authors, with all the information in their Github profile, although most of it is not necessary. Some additional fields could also be added, such as the license type or an abstract.

The version and date-released are optional and could be removed.

Please, check that there is no missing author that should be credited. It is also possible to credit an entity.