LaTeX: partially fix color issues for wrapped code lines with pagebreak
This does not fix entirely #10610 but it does sufficiently for it not to
require reverting #10577 which tried to solve #8686 conundrum.  In
extreme cases, the #8686 problem meant that some contents disappeared
at page bottom, so it is probably better to maintain #10577 which
will avoid any such overflow of code beyond its frame, even though in
some specific cases (a colored entity such as a long string is partly on
both pages), some syntax highlighting gets lost.

There are anyhow other issues with colors for wrapped code lines, even
with no pagebreaks involved, such as #10615.  This patch does not change
the situation there.