Explain non-standard encoding in LatinRules.xdy
In order to fulfill its function, LatinRules.xdy must use single,
non-standard byte characters (neither ASCII, nor multi-byte UTF-8).
To someone encountering the file without knowing its purpose (e.g.
due a post-processing raising a warning for the unrecognized encoding)
this is likely surprising, and may seem like a holdover from a time where
Unicode wasn't as universally supported.

The added comment should make clear that the file must stay as it is,
and in particular that it must not be "standardized" to UTF-8.