always default create optional environment for deprecated builders
Builder implementation would originally always create an optional
environment instance, which an extended builder's implementation could
reference in their own `__init__` call. However, with the change [1] to
support an `env` argument, there is no longer a guarantee that
`self.env` will exist. This commit changes this to restore the creating
of `self.env` if `env` if not passed in.

[1]: 6ef22d261303ec07890305b300135c34952ca327

Signed-off-by: James Knight <james.d.knight@live.com>