Update deploying.rst
Version 0.4 of sphinx-action contains a problem investigated and documented here: https://github.com/ammaraskar/sphinx-action/issues/5. The action throws an error claiming it can't write the log file. The problem is fixed in master and since this is a tutorial, it is more or less expected that it should work off the bat, so I thinking pinning the master version for now and the next release when it comes out, might be the best way to help newbies.