Rename some macros to use no @, make colorgaps true per default
This way, custom user replacement code for
    \sphinxbooktabs{top,mid,bottom}rule
is relatively painless to inject in preamble or via container class.

Also, it was needed to add some LaTeX code to the longtable template in
order to avoid a colour-related problem with \sphinxbottomrule.

The 'sphinxsetup' key now named booktabscolorgaps defaults to being true
and is left only as a way to provide an easy "turn-off" if the issue
with the longtable colours after the mid rule and above the bottom rule
is blocking to some.  But a priori I envision at 6.0.0 to simply remove
this undocumented option as the said issue is not serious and more to be
considered a fact-of-life.  Besides using booktabs+colorrows is strictly
opt-in at this time.  I considered various LaTeX ways around this
longtable thing (it constructs headers and footers once and for all, one
can not let them depend on some "dynamic" context; there is no easy way
for a body row to know it is first on the page after header, and anyway
would require adding mark-up or considerably complicate the \everycr))
but none of them are even remotely reasonable.

Also, a non-documented booktabsnogaps 'sphinxsetup' key has been added:
it suppresses all the extra spacing booktabs injects around its top,
mid, and bottom rules.  Will solve the booktabs+colorrows+longtable
conundrum from previous paragraph.