imgmath: Move dest image file even with embed enabled
We need the generated image files to be in one common
folder to avoid having to re-generate several times the same image
(the sha1 naming of the file acts as a cache to just read the depth)

The location should be possibly common to all workers,
that's why I choosed to just keep the original destination folder,
ie the file is just moved from tempdir to the build folder before.

This significantly drops the build times from 84min to 26min,
but obviously leaves the unused image files in the output folder.

Another solution would be to use another global temporary folder
shared across all the workers, or to add a delete pass after all workers
have joined, but there does not seem to be an api to do that.