apidoc: Compile 'excludes' regexps only once
is_excluded() called fnmatch.fnmatch on each input string, which caused
fnmatch.translate and re.compile for each of them.

This is usually fine because both have their own internal LRU cache,
but when there are more patterns than the cache's size, this causes
every pattern to be recompiled on every call to is_excluded().