Create a search field component
searchfield.html is similar to the existing searchbox.html, but does not
have the heading "Quick Search". Instead, it uses "Search" as a placeholder
text. This gives a cleaner and less cluttered appearance.

The searchfield component is added to the basic theme and thus can be used
in any derivative theme. It is not used by default for backward compatibility.

The searchfield compenent is made the default in the new sphinx13 theme.

CSS adaptions ensure that the top of the search field is aligned with the
top of the main content.