Add faster ``TocTree._toctree_copy`` method (#10988)
As in the standalone html builder the navigation is flattened out for
every single html page, the code needs to create a specialised
toctree for every html page. Previously this was done by deep-copying
the complete navigation toctree and then stripping out the parts not
needed on the particular page. With this change the code only
(deep)-copies the needed parts of the toctree avoiding unnecessary
copying and throwing-away. The performance improvements seems to be
smaller for smaller page counts and get bigger the more pages are
involved.

Co-authored-by: Adam Turner <9087854+AA-Turner@users.noreply.github.com>