Add content_offset parameter to nested_parse_with_titles
Previously, `nested_parse_with_titles` always passed `0` as the input
offset when invoking `nested_parse`.  When parsing the content of a
directive, as is a common use case for `nested_parse_with_titles`,
this leads to incorrect source file/line number information, as it
does not take into account the directive's `content_offset`, which is
always non-zero.

This issue affects *all* object descriptions due to #10887.  It also
affects the `ifconfig` extension.

The `py:module` and `js:module` directives employed a workaround for
this issue, by wrapping the calls to `nested_parse_with_title` with
`switch_source_input`.  That worked, but was more complicated (and
likely less efficient) than necessary.

This commit adds an optional `content_offset` parameter to
`nested_parse_with_titles`, and fixes callers to pass the appropriate
content offset when needed.

This commit eliminates the now-unnecessary calls to
`switch_source_input` and instead specifies the correct `content_offset`.